Handle new case in weights comparison
It is undesirable for the second subgraph in the weights comparison routine (see `compare_weights` method) to have a tag vertex that is not present in the first subgraph, so handle this case by deleting any such vertex and printing a warning message.

Implementing this capability resulted in the addition of new methods (generate_verts_to_edges_mapping and delete_tag_vertex) and a new instance attribute (verts_to_edges) for the Graph class.