Combine as_graph functions
Create a new as_graph function in the new as_graph.py file. This function was created by combining the as_graph functions from convert_json.py and convert_json_sba.py. Because the two implementations shared a lot of code but also deviated in a few important ways, the entire functionality of each was preserved in this function by using the use_sba boolean argument to toggle on/off logical branches according to the implementation in convert_json.py (if False) and the implementation in convert_json_sba.py (if True).

None of the other existing code has been refactored to use this new as_graph function yet.

In this process, the following functions were moved out of graph_utils.py and into the new as_graph.py file:
- pose2diffs
- matrix2measurement
- se3_quat_average