Expand pseudo GT metric validation plotting capabilities
New plotting capability added to the OResultPseudoGTMetricValidation.plot_scatter method: The variable used to plot the colorbar values is now parameterized.

Additionally, replicated documentation in `data_models.py` regarding the need for 'Config' subclasses in certain pydantic models was replaced by a single explanation at the top of the file.