Modified process_graphs.py in run_scripts
process_graphs will now listen to the Firebase for new events when called.
New events will be compared with maps in Firebase for shared cloud anchors.
If shared cloud anchors are found, these maps will be added onto the new event.
This combined map will be optimized and be uploaded back to the Firebase.

Currently, due to processing time, the script limits the number of maps
that can be added on to a new event to five. However, this could be changed.